+++ src/execution/order_index.cpp
@@ -105,8 +105,8 @@ template <class T> int64_t binary_search_lte(uint8_t *data, T key, uint64_t coun
 	bool found = false;
 	auto pos_orig = binary_search(array, key, 0, count, found);
 	assert(pos_orig <= numeric_limits<int32_t>::max());
-	int32_t pos = (int32_t) pos_orig;
-	while (array[pos].value <= key && pos < count)
+	int32_t pos = (int32_t)pos_orig;
+	while (array[pos].value <= key && (uint64_t)pos < count)
 		pos++;
 	return pos;
 }
@@ -116,7 +116,7 @@ template <class T> int64_t binary_search_gte(uint8_t *data, T key, uint64_t coun
 	bool found = false;
 	auto pos_orig = binary_search(array, key, 0, count, found);
 	assert(pos_orig <= numeric_limits<uint32_t>::max());
-	int32_t pos = (int32_t) pos_orig;
+	int32_t pos = (int32_t)pos_orig;
 	if (found) {
 		while (pos >= 0 && array[pos].value == key)
 			pos--;
