+++ src/planner/expression_binder/order_binder.cpp
@@ -38,7 +38,7 @@ unique_ptr<Expression> OrderBinder::Bind(unique_ptr<ParsedExpression> expr) {
 		}
 		// INTEGER constant: we use the integer as an index into the select list (e.g. ORDER BY 1)
 		auto index = constant.value.GetNumericValue();
-		if (index < 1 || index > node.select_list.size()) {
+		if (index < 1 || (uint64_t)index > node.select_list.size()) {
 			throw BinderException("ORDER term out of range - should be between 1 and %d", (int)node.select_list.size());
 		}
 		return CreateProjectionReference(*expr, index - 1);
